+++ mypy/checkexpr.py
@@ -21,7 +21,7 @@ from mypy.types import (
     PartialType, DeletedType, UninhabitedType, TypeType, TypeOfAny, LiteralType, LiteralValue,
     is_named_instance, FunctionLike, ParamSpecType, ParamSpecFlavor,
     StarType, is_optional, remove_optional, is_generic_instance, get_proper_type, ProperType,
-    get_proper_types, flatten_nested_unions, LITERAL_TYPE_NAMES,
+    get_proper_types, flatten_nested_unions, LITERAL_TYPE_NAMES, UntypedType,
 )
 from mypy.nodes import (
     AssertTypeExpr, NameExpr, RefExpr, Var, FuncDef, OverloadedFuncDef, TypeInfo, CallExpr,
@@ -358,21 +358,59 @@ class ExpressionChecker(ExpressionVisitor[Type]):
                 isinstance(callee_type, CallableType)):
             if callee_type.implicit:
                 self.msg.untyped_function_call(callee_type, e)
-            elif not callee_type.fully_typed:
-                # check for partially typed defs
-                if isinstance(callee_type.definition, FuncDef):
-                    if callee_type.definition.info:
-                        callee_module = callee_type.definition.info.module_name
+            elif has_untyped_type(callee_type):
+                # Get module of the function, to get its settings
+                #  This is fairly sus, and I'm sure there are cases where it gets
+                #   the wrong module. We should instead set it in CallableType
+                if isinstance(e.callee, MemberExpr):
+                    if e.callee.kind is None:
+                        # class, kinda hacky
+                        it = get_proper_type(self.chk._type_maps[0][e.callee.expr])
+                        if isinstance(it, CallableType):
+                            # callable class reference
+                            ret_type_ = get_proper_type(it.ret_type)
+                            if isinstance(ret_type_, TupleType):
+                                callee_module = ret_type_.partial_fallback.type.module_name
+                            elif isinstance(ret_type_, Instance):
+                                callee_module = ret_type_.type.module_name
+                            else:
+                                callee_module = None
+                        elif isinstance(it, TypeType):
+                            # type reference
+                            if isinstance(it.item, TupleType):
+                                callee_module = it.item.partial_fallback.type.module_name
+                            elif isinstance(it.item, Instance):
+                                callee_module = it.item.type.module_name
+                            elif isinstance(it.item, UnionType):
+                                # TODO: have to figure out which part of the union has the Untyped
+                                callee_module = self.chk.tree.name
+                            else:
+                                callee_module = None
+                        elif isinstance(it, Instance):
+                            # instance reference
+                            callee_module = it.type.module_name
+                        elif isinstance(it, UnionType):
+                            # TODO: have to figure out which part of the union has the Untyped
+                            callee_module = self.chk.tree.name
+                        else:
+                            callee_module = None
                     else:
-                        callee_module = callee_type.definition.fullname.rpartition(".")[0]
+                        # module
+                        assert isinstance(e.callee.expr, RefExpr)
+                        callee_module = e.callee.expr.fullname
                 elif isinstance(e.callee, NameExpr):
                     assert e.callee.fullname
-                    callee_module = e.callee.fullname.rpartition(".")[0]
+                    if "." not in e.callee.fullname:
+                        # local def
+                        callee_module = self.chk.tree.name
+                    else:
+                        callee_module = e.callee.fullname.rpartition(".")[0]
                 elif isinstance(e.callee, MemberExpr) and isinstance(e.callee.expr, NameExpr):
                     assert e.callee.expr.fullname
                     callee_module = e.callee.expr.fullname.rpartition(".")[0]
                 else:
                     # If this branch gets hit then look for a new way to get the module name
+                    # TODO: test time error
                     callee_module = None
                 if callee_module:
                     callee_options = self.chk.options.per_module(callee_module)
@@ -1323,7 +1361,7 @@ class ExpressionChecker(ExpressionVisitor[Type]):
         else:
             # In dynamically typed functions use implicit 'Any' types for
             # type variables.
-            inferred_args = [AnyType(TypeOfAny.unannotated)] * len(callee_type.variables)
+            inferred_args = [UntypedType()] * len(callee_type.variables)
         return self.apply_inferred_arguments(callee_type, inferred_args,
                                              context)
 
@@ -2065,7 +2103,9 @@ class ExpressionChecker(ExpressionVisitor[Type]):
     def check_any_type_call(self, args: List[Expression], callee: Type) -> Tuple[Type, Type]:
         self.infer_arg_types_in_empty_context(args)
         callee = get_proper_type(callee)
-        if isinstance(callee, AnyType):
+        if isinstance(callee, UntypedType):
+            return UntypedType(), UntypedType()
+        elif isinstance(callee, AnyType):
             return (AnyType(TypeOfAny.from_another_any, source_any=callee),
                     AnyType(TypeOfAny.from_another_any, source_any=callee))
         else:
@@ -3729,7 +3769,7 @@ class ExpressionChecker(ExpressionVisitor[Type]):
         """
 
         if not self.chk.in_checked_function():
-            return AnyType(TypeOfAny.unannotated)
+            return UntypedType()
         elif len(e.call.args) == 0:
             if self.chk.options.python_version[0] == 2:
                 self.chk.fail(message_registry.TOO_FEW_ARGS_FOR_SUPER, e)
@@ -4034,6 +4074,7 @@ class ExpressionChecker(ExpressionVisitor[Type]):
             self.msg.disallowed_any_type(typ, node)
 
         if not self.chk.in_checked_function() or self.chk.current_node_deferred:
+            # this is more of a placeholder I think
             return AnyType(TypeOfAny.unannotated)
         else:
             return typ
@@ -4319,6 +4360,11 @@ def has_any_type(
     return t.accept(HasAnyType(ignore_in_type_obj, ignore_any_from_error=ignore_any_from_error))
 
 
+def has_untyped_type(t: Type, ignore_in_type_obj: bool = False) -> bool:
+    """Whether ``t`` contains an untyped type"""
+    return t.accept(HasUntypedType())
+
+
 class HasAnyType(types.TypeQuery[bool]):
     def __init__(self, ignore_in_type_obj: bool, ignore_any_from_error: bool = False) -> None:
         super().__init__(any)
@@ -4338,6 +4384,15 @@ class HasAnyType(types.TypeQuery[bool]):
         if self.ignore_in_type_obj and t.is_type_obj():
             return False
         return super().visit_callable_type(t)
+
+
+class HasUntypedType(types.TypeQuery[bool]):
+    def __init__(self) -> None:
+        super().__init__(any)
+
+    def visit_any(self, t: AnyType) -> bool:
+        return isinstance(t, UntypedType) or t.type_of_any in (
+            TypeOfAny.unannotated, TypeOfAny.from_omitted_generics)
 
 
 def has_coroutine_decorator(t: Type) -> bool: