+++ mypy/test/data.py
@@ -579,7 +579,7 @@ def pytest_pycollect_makeitem(collector: Any, name: str,
             # Non-None result means this obj is a test case.
             # The collect method of the returned DataSuiteCollector instance will be called later,
             # with self.obj being obj.
-            return DataSuiteCollector.from_parent(  # type: ignore[no-untyped-call]
+            return DataSuiteCollector.from_parent(  # type: ignore[unused-ignore, no-untyped-call]
                 parent=collector, name=name,
             )
     return None
@@ -614,7 +614,7 @@ def split_test_cases(parent: 'DataFileCollector', suite: 'DataSuite',
                 name, parent.name, line_no,
             ))
         platform = platform_flag[1:] if platform_flag else None
-        yield DataDrivenTestCase.from_parent(
+        yield DataDrivenTestCase.from_parent(  # type: ignore[unused-ignore, no-untyped-call]
             parent=parent,
             suite=suite,
             file=file,